fixed cleanDB.ts, added main.workflow.yml and dev.workflow.yml

+++ server/src/seeds/cleanDb.ts
@@ -3,12 +3,20 @@ import db from "../config/connection.js";
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists =
-      (await models[modelName]?.db?.db
-        ?.listCollections({
-          name: collectionName,
-        })
-        ?.toArray()) || [];
+    if (
+      !models ||
+      !models[modelName] ||
+      !models[modelName].db ||
+      !models[modelName].db.db
+    ) {
+      throw new Error(`Model ${modelName} does not exist`);
+    }
+
+    let modelExists = await models[modelName].db.db
+      .listCollections({
+        name: collectionName,
+      })
+      .toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
